Move handler de api para arquivo central que gerencia handlers de requests não-2xx
