+++ traffic_control/handlers.py
@@ -4,6 +4,8 @@ import random
 
 from django.shortcuts import render
 from ratelimit.exceptions import Ratelimited
+from rest_framework.exceptions import Throttled
+from rest_framework.views import exception_handler
 
 from traffic_control.logging import log_blocked_request
 
@@ -11,6 +13,12 @@ rate_limit_msg = """
 <p>Você atingiu o limite de requisições e, por isso, essa requisição foi bloqueada. Caso você precise acessar várias páginas de um dataset, por favor, baixe o dataset completo em vez de percorrer várias páginas na interface (o link para baixar o arquivo completo encontra-se na <a href="https://brasil.io/datasets/">página do dataset</a>).</p>
 <p>Utilizar a interface do Brasil.io via web crawlers e de maneira não otimizada onera muito nossos servidores e atrapalha a experiência de outros usuários. Se o abuso continuar, precisaremos restringir ainda mais os limites de requisições e não gostaríamos de fazer isso.</p>
 <p>Lembre-se: o Brasil.IO é um projeto colaborativo, desenvolvido por voluntários e mantido por financiamento coletivo, você pode doar na <a href="https://apoia.se/brasilio">página do projeto no Apoia.se</a>.</p>
+""".strip()
+
+api_throtthling_msg = """
+Você atingiu o limite de requisições e, por isso, essa requisição foi bloqueada. Caso você precise acessar várias páginas de um dataset, por favor, baixe o dataset completo em vez de percorrer várias páginas na API (o link para baixar o arquivo completo encontra-se na página do dataset, em https://brasil.io/datasets/).
+Utilizar a API desnecessariamente e de maneira não otimizada onera muito nossos servidores e atrapalha a experiência de outros usuários. Se o abuso continuar, precisaremos restringir ainda mais a API e não gostaríamos de fazer isso.
+Lembre-se: o Brasil.IO é um projeto colaborativo, desenvolvido por voluntários e mantido por financiamento coletivo, você pode doar para o projeto em: https://apoia.se/brasilio
 """.strip()
 
 
@@ -30,3 +38,17 @@ def handler_403(request, exception):
     msg += "<!-- " + data + " -->"
     context = {"title_4xx": status, "message": msg}
     return render(request, "4xx.html", context, status=status)
+
+
+def api_exception_handler(exc, context):
+    response = exception_handler(exc, context)
+    status_code = getattr(response, "status_code", None)
+
+    if isinstance(exc, Throttled):
+        custom_response_data = {"message": api_throtthling_msg, "available_in": f"{exc.wait} seconds"}
+        response.data = custom_response_data
+
+    if 400 <= status_code < 500:
+        log_blocked_request(context["request"], status_code)
+
+    return response
